3.0.1 . local ver working! without indexer

+++ src/ReadFile.java
@@ -3,6 +3,7 @@ import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
+import sun.awt.Mutex;
 
 
 import java.io.*;
@@ -10,6 +11,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.*;
+import java.util.concurrent.ConcurrentHashMap;
 
 import static java.nio.file.StandardOpenOption.APPEND;
 import static java.nio.file.StandardOpenOption.CREATE;
@@ -17,28 +19,34 @@ import static java.nio.file.StandardOpenOption.CREATE;
 public class ReadFile implements Runnable{
 
     private Path pathData;
-    private volatile Integer fileNumber ;
+    private static volatile Integer fileNumber = 1 ;
     private Parse parser;
     private File folder ;
-    private volatile HashSet<String> filesExecuted ;
-    private volatile Object lookIncrease;
+
+    private static volatile ConcurrentHashMap<String,Integer> filesExecuted = new ConcurrentHashMap<>();
+    private static volatile Object lookIncrease  = new Object();
+    private static volatile Object lookClean  = new Object();
     private Map<String , Map<Integer,Integer>> Voc ;
     private Map<Integer , String> fileMeta_data;
+    private static volatile boolean cleanfile = true;
     private int fileIteration ;
 
     public ReadFile(Path pathData , final File folder) {
-        if (Files.exists(pathData)) {
-            // file exist
-            System.out.println("exist");
-            deleteDirectory(new File(pathData.toString()));
-        }
+            synchronized (lookClean) {
+                if (cleanfile) {
+                    cleanfile = false;
+                    if (Files.exists(pathData)) {
+                        // file exist
+                        System.out.println("exist");
+                        deleteDirectory(new File(pathData.toString()));
+                    }
 
+                }
+
+            }
         new File(pathData.toString()).mkdirs();
         this.pathData = pathData;
-        fileNumber=0;
-        this.folder=folder;
-        this.filesExecuted = new HashSet<>();
-        lookIncrease = new Object();
+        this.folder = folder;
     }
 
     private boolean deleteDirectory(File directoryToBeDeleted) {
@@ -57,9 +65,9 @@ public class ReadFile implements Runnable{
                 listFilesForFolder(fileEntry);
             } else {
                 boolean check = false ;
-                synchronized (this) {
-                    if (!filesExecuted.contains(fileEntry.toString())) {
-                        filesExecuted.add(fileEntry.toString());
+                synchronized (lookClean) {
+                    if (!filesExecuted.containsKey(fileEntry.toPath().toString())) {
+                        filesExecuted.put(fileEntry.toPath().toString(),0);
                         check = true ;
                     }
                 }
@@ -72,11 +80,7 @@ public class ReadFile implements Runnable{
     }
 
     public void readFile(Path file ,String fileName) {
-//        try (InputStream in = Files.newInputStream(file);
-//             BufferedReader reader =
-//                     new BufferedReader(new InputStreamReader(in))) {
-//            String line = null;
-//            String text="" ;
+
         Voc = new HashMap<>();
         try {
                   Document document = Jsoup.parse(new String(Files.readAllBytes(file)));
@@ -92,20 +96,25 @@ public class ReadFile implements Runnable{
                 Elements title =  doc.getElementsByTag("TI");
                 Elements date =  doc.getElementsByTag("DATE1");
                 Elements docno =  doc.getElementsByTag("DOCNO");
-                fileMeta_data.put(local_file_num , "="+title.text()+"="+date.text()+"="+docno.text());
+
 
                 Map<String,Integer> mp = parser.parseIt(doc.text());
-                Iterator<String> it = mp.keySet().iterator();
-                while (it.hasNext()){
-                    String s = it.next();
+                int max_tf = 0 ;
+                String[] it = mp.keySet().toArray(new String[mp.size()]);
+                for (String s:
+                        it) {
                     if(Voc.get(s)==null){
                         Map<Integer,Integer>  l= new HashMap<>();
                         Voc.put(s , l);
                     }
                     Voc.get(s).put(local_file_num , mp.get(s));
+                    if(mp.get(s)>max_tf){
+                        max_tf=mp.get(s);
+                    }
                 }
+                fileMeta_data.put(local_file_num , "|"+title.text()+"|"+date.text()+"|"+docno.text()+"|max_tf|"+max_tf +"|qw|"+mp.size());
                 fileIteration++;
-                if(fileIteration>1000){
+                if(fileIteration>=1000){
                     fileIteration = 0 ;
                     writeFile(fileMeta_data.toString() , local_file_num.toString());
                     writeFile(Voc.toString() , local_file_num.toString()+"Voc");
